Change User ID to Email for Project Member Addition
In this commit, the key identifier for adding a project member switched from User ID to Email. This change is implemented in AddMemberCommand.cs and its Validator.cs. Moreover, email validation is enforced to ensure its correctness and existence. This alteration aims to enhance the user-friendliness of our application since it is easier for users to recall their emails compared to User IDs. It also checks for the user's existence before project member addition, preventing potential errors.