Update task assignment method in BaseProjectTask
Updated the task assignment method in BaseProjectTask to include the user performing the assignment and prevent an event from being triggered when the assigner and assignee are the same user or the assignee is null. This is to improve user experience, eliminate redundant actions and optimize the task assignment process.